Watts Max Calculation
Created a basic integration for watts max. Needs a more sophisticated approach for calculating the maximum watts. Currently uses a test case based on the second maximum value.

Need to create a function to keep testing the value based off of if the original test case works. I could also just eliminate all of the outliers.