Overhaul To WattsMax
Finally found a good method of determining the watts max using Dr. Cooper's criteria.

Using for loops, counters, and if statements, I parse through all data and determine the watts max and make sure that this max value is not an outlier of the data set.

I have tested this using test cases as well.