Roll Mean / R^2 Analysis Final
Fixed all of the data to roll means. All of the data is already roll meaned upon cleaning. No need to manually mean the data before hand.

The problem with the mean method is that I deleted 4 datapoints from the graph that are stored as NA when the roll mean is calculated. We will work on elaborating on that later.

 I also replaced the first watts structure with the linearity machine implementation.