USER INTERFACE OVERHAUL
Converted the entire userinterface into the shiny dashboard user interface and it looks amazing. This framework primarily relies on the use of boxes which makes the dashboard very nice.

As of right now I got rid of the entire test validity graphs because we do not need them in the future application.

I added CSS tags in order to create in line elements for the shiny application. (Top demographic information)