Overhaul to Table Implementation
Added new functionality to the gt table that is under the tab that has all of the reference values.

So basically, I created functions that calculated all of the reference values and incorporated them into this table as well.  I did not use data abstraction for this and should do this in the future at some point.

With this data, I inputted it into the GT table using a data frame. In the future, I should convert this into tibble format for ease of use.